Added support for bigger numbers

+++ NumberImput/voids.cpp
@@ -5,14 +5,14 @@
 
 void printStuff()
 {
-	int x = userNumber();
+	int64_t x = userNumber();
 	std::cout << x << " doubled is:" << doubleNumber(x) << std::endl;
 	printStuff1();
 }
 
 void printStuff1()
 {
-	int x = userNumber();
+	int64_t x = userNumber();
 	std::cout << x << " doubled is:" << doubleNumber(x) << std::endl;
 	printStuff();
 }
